+++ example/src/commonMain/kotlin/com/usmonie/compass/example/LoginScreen.kt
@@ -1,7 +1,13 @@
 package com.usmonie.compass.example
 
-import User
 import com.usmonie.compass.screen.state.navigation.ScreenId
+import kotlinx.serialization.Serializable
 
+@Serializable
 internal object LoginScreen : ScreenId("login")
+@Serializable
 internal data class ProfileScreen(val user: User) : ScreenId("profile")
+
+@Serializable
+internal object CounterScreen : ScreenId("counter")
+