+++ example/src/commonMain/kotlin/com/usmonie/compass/example/User.kt
@@ -1,6 +1,10 @@
+package com.usmonie.compass.example
+
+import kotlinx.serialization.Serializable
+
+@Serializable
 internal data class User(val id: String, val name: String)
 
-// model/ApiClient.kt
 internal object ApiClient {
     suspend fun fetchUser(userId: String): User {
         // Simulate network delay